D: 优化一帧的Tick逻辑， 优化掉等待的部分，
1. 优化掉MainThread和 RenderThread的等待逻辑。 由SCSP有数据判定RenderThread是否启动
2. MainThread只等待mFrameContext可以复用，FrameContext记录完成的GPUFenceValue
3. RenderThread提供3个CommandAlloc，避免重置时GPU仍在运行（CommandAlloc需要等GPU执行完才能Reset）

+++ Runtime/Platforms/D3D12/D3D12Struct.h
@@ -26,7 +26,7 @@ namespace EngineCore
     struct TD3D12Fence
     {
     public:
-        int mCurrentFence;
+        uint64_t mCurrentFence;
         Microsoft::WRL::ComPtr<ID3D12Fence> mFence;
     };
